requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Neuroimaging
nibabel>=5.1.0
nilearn>=0.10.0
SimpleITK>=2.2.0

# Vision Transformer and Deep Learning
einops>=0.6.0
timm>=0.9.0
torch-geometric>=2.3.0

# Privacy-Preserving ML
tenseal>=0.3.14
opacus>=1.4.0
syft>=0.8.0

# Explainability
lime>=0.2.0
shap>=0.42.0
captum>=0.6.0

# Federated Learning
flwr>=1.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.4.0
